+++ src/main/java/io/myzticbean/finditemaddon/utils/LocationUtils.java
@@ -22,7 +22,6 @@ import io.myzticbean.finditemaddon.FindItemAddOn;
 import io.myzticbean.finditemaddon.models.enums.PlayerPermsEnum;
 import io.myzticbean.finditemaddon.utils.log.Logger;
 import lombok.experimental.UtilityClass;
-import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.World;
@@ -32,6 +31,9 @@ import org.jetbrains.annotations.Nullable;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.HashSet;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 /**
  * @author myzticbean
@@ -39,11 +41,16 @@ import java.util.List;
 @UtilityClass
 public class LocationUtils {
 
-    private static final List<Material> damagingBlocks = new ArrayList<>();
-    private static final List<Material> nonSuffocatingBlocks = new ArrayList<>();
+    private static final Set<Material> damagingBlocks = new HashSet<>();
+    private static final Set<Material> nonSuffocatingBlocks = new HashSet<>();
+    private static final Set<Material> walledSignBlocks = new HashSet<>();
     private static final int BELOW_SAFE_BLOCK_CHECK_LIMIT = 20;
 
     static {
+        walledSignBlocks.addAll(Arrays.stream(Material.values())
+                .filter(p -> p.toString().toLowerCase().contains("_wall_sign"))
+                .collect(Collectors.toSet())
+        );
         // Initializing Damaging blocks
         damagingBlocks.add(Material.LAVA);
         damagingBlocks.add(Material.CACTUS);
@@ -61,11 +68,15 @@ public class LocationUtils {
         // Glass
         nonSuffocatingBlocks.add(Material.GLASS);
         nonSuffocatingBlocks.addAll(
-                new ArrayList<>(Arrays.stream(Material.values()).filter(p -> p.toString().toLowerCase().contains("_glass")).toList())
+                Arrays.stream(Material.values())
+                        .filter(p -> p.toString().toLowerCase().contains("_glass"))
+                        .collect(Collectors.toSet())
         );
         // Glass Panes
         nonSuffocatingBlocks.addAll(
-                new ArrayList<>(Arrays.stream(Material.values()).filter(p -> p.toString().toLowerCase().contains("glass_pane")).toList())
+                Arrays.stream(Material.values())
+                        .filter(p -> p.toString().toLowerCase().contains("glass_pane"))
+                        .collect(Collectors.toSet())
         );
         // Leaves
         nonSuffocatingBlocks.add(Material.ACACIA_LEAVES);
@@ -74,21 +85,33 @@ public class LocationUtils {
         nonSuffocatingBlocks.add(Material.JUNGLE_LEAVES);
         nonSuffocatingBlocks.add(Material.OAK_LEAVES);
         nonSuffocatingBlocks.add(Material.SPRUCE_LEAVES);
-        if(Bukkit.getServer().getVersion().contains("1.17")) {
-            nonSuffocatingBlocks.add(Material.AZALEA_LEAVES);
-            nonSuffocatingBlocks.add(Material.FLOWERING_AZALEA_LEAVES);
-        }
+        nonSuffocatingBlocks.add(Material.AZALEA_LEAVES);
+        nonSuffocatingBlocks.add(Material.FLOWERING_AZALEA_LEAVES);
         // Slabs
         nonSuffocatingBlocks.addAll(
-                new ArrayList<>(Arrays.stream(Material.values()).filter(p -> p.toString().toLowerCase().contains("_slab")).toList())
+                Arrays.stream(Material.values())
+                        .filter(p -> p.toString().toLowerCase().contains("_slab"))
+                        .collect(Collectors.toSet())
         );
         // Walled Signs
+        nonSuffocatingBlocks.addAll(walledSignBlocks);
+        // Hanging Signs
         nonSuffocatingBlocks.addAll(
-                new ArrayList<>(Arrays.stream(Material.values()).filter(p -> p.toString().toLowerCase().contains("_sign")).toList())
+                Arrays.stream(Material.values())
+                        .filter(p -> p.toString().toLowerCase().contains("_hanging_sign"))
+                        .collect(Collectors.toSet())
         );
         // Stairs
         nonSuffocatingBlocks.addAll(
-                new ArrayList<>(Arrays.stream(Material.values()).filter(p -> p.toString().toLowerCase().contains("_stairs")).toList())
+                Arrays.stream(Material.values())
+                        .filter(p -> p.toString().toLowerCase().contains("_stairs"))
+                        .collect(Collectors.toSet())
+        );
+        // Fences
+        nonSuffocatingBlocks.addAll(
+                Arrays.stream(Material.values())
+                        .filter(p -> p.toString().toLowerCase().contains("_fence"))
+                        .collect(Collectors.toSet())
         );
         // MISC
         nonSuffocatingBlocks.add(Material.HONEY_BLOCK);
@@ -99,14 +122,13 @@ public class LocationUtils {
         nonSuffocatingBlocks.add(Material.COMPOSTER);
         nonSuffocatingBlocks.add(Material.GRINDSTONE);
         nonSuffocatingBlocks.add(Material.STONECUTTER);
-        nonSuffocatingBlocks.add(Material.IRON_BARS);
         nonSuffocatingBlocks.add(Material.END_PORTAL_FRAME);
         nonSuffocatingBlocks.add(Material.PISTON_HEAD);
-
     }
 
     @Nullable
     public static Location findSafeLocationAroundShop(Location shopLocation, Player player) {
+        Logger.logDebugInfo("Finding safe location around the shop");
         Location roundedShopLoc = getRoundedDestination(shopLocation);
         Logger.logDebugInfo("Rounded location: " + roundedShopLoc.getX() + ", " + roundedShopLoc.getY() + ", " + roundedShopLoc.getZ());
         // Creating a list of four block locations in 4 sides of the shop
@@ -135,43 +157,44 @@ public class LocationUtils {
                 roundedShopLoc.getY(),
                 roundedShopLoc.getZ() - 1
         ));
-        for(Location loc_i : possibleSafeLocList) {
-            Logger.logDebugInfo("Possible safe location: " + loc_i.getX() + ", " + loc_i.getY() + ", " + loc_i.getZ());
-            if(loc_i.getBlock().getType().equals(FindItemAddOn.getQsApiInstance().getShopSignMaterial())) {
-                Logger.logDebugInfo("Shop sign block found at " + loc_i.getX() + ", " + loc_i.getY() + ", " + loc_i.getZ());
+        for(Location possibleSafeLoc : possibleSafeLocList) {
+            Logger.logDebugInfo("Possible safe location: " + possibleSafeLoc.getX() + ", " + possibleSafeLoc.getY() + ", " + possibleSafeLoc.getZ());
+            if(possibleSafeLoc.getBlock().getType().equals(FindItemAddOn.getQsApiInstance().getShopSignMaterial())
+                || walledSignBlocks.contains(possibleSafeLoc.getBlock().getType())) {
+                Logger.logDebugInfo("Shop sign block found at " + possibleSafeLoc.getX() + ", " + possibleSafeLoc.getY() + ", " + possibleSafeLoc.getZ());
                 // Adding a check for a safe location check bypass permission
-                if(player.hasPermission(PlayerPermsEnum.FINDITEM_SHOPTP_BYPASS_SAFETYCHECK.value())) {
+                if(PlayerUtil.hasPermission(player, PlayerPermsEnum.FINDITEM_SHOPTP_BYPASS_SAFETYCHECK.value())) {
                     Location blockBelow = new Location(
-                            loc_i.getWorld(),
-                            loc_i.getBlockX(),
-                            loc_i.getBlockY() - 1,
-                            loc_i.getBlockZ()
+                            possibleSafeLoc.getWorld(),
+                            possibleSafeLoc.getBlockX(),
+                            possibleSafeLoc.getBlockY() - 1,
+                            possibleSafeLoc.getBlockZ()
                     );
-                    loc_i = lookAt(getRoundedDestination(new Location(
-                            blockBelow.getWorld(),
-                            blockBelow.getX(),
-                            blockBelow.getY() + 1,
-                            blockBelow.getZ()
-                    )), roundedShopLoc);
-                    return loc_i;
+                    possibleSafeLoc = lookAt(getRoundedDestination(new Location(
+                                blockBelow.getWorld(),
+                                blockBelow.getX(),
+                                blockBelow.getY() + 1,
+                                blockBelow.getZ()
+                        )
+                    ), roundedShopLoc);
+                    return possibleSafeLoc;
                 }
                 // check if the block above is suffocating
                 Location blockAbove = new Location(
-                        loc_i.getWorld(),
-                        loc_i.getBlockX(),
-                        loc_i.getBlockY() + 1,
-                        loc_i.getBlockZ());
-                Logger.logDebugInfo("Block above shop sign: "
-                        + blockAbove.getX() + ", " + blockAbove.getY() + ", " + blockAbove.getZ());
+                        possibleSafeLoc.getWorld(),
+                        possibleSafeLoc.getBlockX(),
+                        possibleSafeLoc.getBlockY() + 1,
+                        possibleSafeLoc.getBlockZ());
+                Logger.logDebugInfo("Block above shop sign: " + blockAbove.getX() + ", " + blockAbove.getY() + ", " + blockAbove.getZ());
                 if(!isBlockSuffocating(blockAbove)) {
                     Location blockBelow = null;
                     boolean safeLocFound = false;
                     for(int i = 1; i <= BELOW_SAFE_BLOCK_CHECK_LIMIT; i++) {
                         blockBelow = new Location(
-                                loc_i.getWorld(),
-                                loc_i.getBlockX(),
-                                loc_i.getBlockY() - i,
-                                loc_i.getBlockZ()
+                                possibleSafeLoc.getWorld(),
+                                possibleSafeLoc.getBlockX(),
+                                possibleSafeLoc.getBlockY() - i,
+                                possibleSafeLoc.getBlockZ()
                         );
                         Logger.logDebugInfo("Block below shop sign: "
                                 + blockBelow.getBlock().getType() + " " + blockBelow.getX() + ", " + blockBelow.getY() + ", " + blockBelow.getZ());
@@ -192,13 +215,15 @@ public class LocationUtils {
                         }
                     }
                     if(safeLocFound) {
-                        loc_i = lookAt(getRoundedDestination(new Location(
-                                blockBelow.getWorld(),
-                                blockBelow.getX(),
-                                blockBelow.getY() + 1,
-                                blockBelow.getZ()
-                        )), roundedShopLoc);
-                        return loc_i;
+                        possibleSafeLoc = lookAt(getRoundedDestination(new Location(
+                                        blockBelow.getWorld(),
+                                        blockBelow.getX(),
+                                        blockBelow.getY() + 1,
+                                        blockBelow.getZ()
+                                )
+                        ), roundedShopLoc);
+                        Logger.logDebugInfo("Safe location found: " + possibleSafeLoc.getX() + ", " + possibleSafeLoc.getY() + ", " + possibleSafeLoc.getZ());
+                        return possibleSafeLoc;
                     }
                     else {
                         return null;
@@ -210,7 +235,7 @@ public class LocationUtils {
                 }
             }
             else {
-                Logger.logDebugInfo("Block not shop sign. Block type: " + loc_i.getBlock().getType());
+                Logger.logDebugInfo("Block not shop sign. Block type: " + possibleSafeLoc.getBlock().getType());
             }
         }
         Logger.logDebugInfo("No safe block found near shop");
@@ -218,14 +243,14 @@ public class LocationUtils {
     }
 
     // Found the below function from this thread: https://bukkit.org/threads/lookat-and-move-functions.26768/
-    private static Location lookAt(Location loc, Location lookat) {
+    private static Location lookAt(Location loc, Location lookAt) {
         //Clone the loc to prevent applied changes to the input loc
         loc = loc.clone();
 
         // Values of change in distance (make it relative)
-        double dx = lookat.getX() - loc.getX();
-        double dy = lookat.getY() - loc.getY();
-        double dz = lookat.getZ() - loc.getZ();
+        double dx = lookAt.getX() - loc.getX();
+        double dy = lookAt.getY() - loc.getY();
+        double dz = lookAt.getZ() - loc.getZ();
 
         // Set yaw
         if (dx != 0) {